refactor: generic extension crate to de-duplicate code between runtimes (#163)
