+++ Cargo.toml
@@ -23,18 +23,19 @@ members = [
     "pallets/*",
     "primitives",
 ]
-exclude = [
-    "pop-api",
-    "tests/contracts"
-]
+exclude = ["pop-api", "tests/contracts"]
 
 resolver = "2"
 
 [workspace.dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = [
+    "derive",
+] }
 hex-literal = "0.4.1"
 log = { version = "0.4.20", default-features = false }
-scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
+scale-info = { version = "2.10.0", default-features = false, features = [
+    "derive",
+] }
 smallvec = "1.11.0"
 serde = "1.0.195"
 clap = { version = "4.4.18", features = ["derive"] }
@@ -52,7 +53,8 @@ substrate-build-script-utils = "11.0.0"
 
 # Local
 pallet-api = { path = "pallets/api", default-features = false }
-pop-runtime-devnet = { path = "runtime/devnet", default-features = true } # default-features=true required for `-p pop-node` builds
+pop-chain-extension = { path = "./extension", default-features = false }
+pop-runtime-devnet = { path = "runtime/devnet", default-features = true }   # default-features=true required for `-p pop-node` builds
 pop-runtime-testnet = { path = "runtime/testnet", default-features = true } # default-features=true required for `-p pop-node` builds
 pop-runtime-common = { path = "runtime/common", default-features = false }
 pop-primitives = { path = "./primitives", default-features = false }
@@ -82,7 +84,9 @@ frame-system = { version = "29.0.0", default-features = false }
 frame-system-benchmarking = { version = "29.0.0", default-features = false }
 frame-system-rpc-runtime-api = { version = "27.0.0", default-features = false }
 frame-try-runtime = { version = "0.35.0", default-features = false }
-pallet-aura = { version = "28.0.0", default-features = false, features = ["experimental"] }
+pallet-aura = { version = "28.0.0", default-features = false, features = [
+    "experimental",
+] }
 pallet-authorship = { version = "29.0.0", default-features = false }
 pallet-assets = { version = "30.0.0", default-features = false }
 pallet-balances = { version = "29.0.2", default-features = false }
@@ -139,7 +143,9 @@ xcm-executor = { package = "staging-xcm-executor", version = "8.0.2", default-fe
 # Cumulus
 asset-test-utils = { version = "8.0.1", default-features = false }
 cumulus-pallet-aura-ext = { version = "0.8.0", default-features = false }
-cumulus-pallet-parachain-system = { version = "0.8.1", default-features = false, features = ["parameterized-consensus-hook"] }
+cumulus-pallet-parachain-system = { version = "0.8.1", default-features = false, features = [
+    "parameterized-consensus-hook",
+] }
 cumulus-pallet-session-benchmarking = { version = "10.0.0", default-features = false }
 cumulus-pallet-xcm = { version = "0.8.0", default-features = false }
 cumulus-pallet-xcmp-queue = { version = "0.8.0", default-features = false }
