+++ extension/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "pop-chain-extension"
+version = "0.1.0"
+authors.workspace = true
+description.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+edition.workspace = true
+publish = false
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+codec.workspace = true
+log.workspace = true
+
+# Local
+pop-primitives.workspace = true
+
+# Substrate
+frame-support.workspace = true
+frame-system.workspace = true
+pallet-contracts.workspace = true
+sp-core.workspace = true
+sp-runtime.workspace = true
+sp-std.workspace = true
+
+[dev-dependencies]
+rand = "0.8.5"
+
+[features]
+default = ["std"]
+std = [
+ "log/std",
+ "codec/std",
+ "frame-support/std",
+ "frame-system/std",
+ "pallet-contracts/std",
+ "pop-primitives/std",
+ "sp-runtime/std",
+ "sp-core/std",
+ "sp-std/std",
+]
+runtime-benchmarks = [
+ "frame-support/runtime-benchmarks",
+ "frame-system/runtime-benchmarks",
+ "pallet-contracts/runtime-benchmarks",
+ "sp-runtime/runtime-benchmarks",
+]
