+++ runtime/devnet/Cargo.toml
@@ -22,6 +22,7 @@ scale-info.workspace = true
 smallvec.workspace = true
 
 # Local
+pop-chain-extension.workspace = true
 pop-primitives.workspace = true
 pop-runtime-common.workspace = true
 pallet-api.workspace = true
@@ -139,6 +140,7 @@ std = [
     "parachains-common/std",
     "polkadot-parachain-primitives/std",
     "polkadot-runtime-common/std",
+    "pop-chain-extension/std",
     "pop-primitives/std",
     "scale-info/std",
     "sp-api/std",
